+++ docs/changes.rst
@@ -1,10 +1,27 @@
 Changes
 =======
 
-Version X.X
------------
+Version 1.3.0
+-------------
 
-Documentation improvements by :user:`gamesbook` (:issue:`458`).
+* Pass through keyword arguments to :func:`xlrd.open_workbook`. By :user:`gjunqueira`,
+  :issue:`470`, :issue:`473`.
+
+* Fixes for upstream API changes in openpyxl and intervaltree modules. The arguments
+  to :func:`petl.io.xlsx.fromxlsx` have changed for specifying row and column offsets
+  to match openpyxl. (:issue:`472` - :user:`alimanfoo`).
+  
+* Exposed `read_only` argument in :func:`petl.io.xlsx.fromxlsx` and set default to 
+  False to prevent truncation of files created by LibreOffice. By :user:`mbelmadani`, 
+  :issue:`457`.
+
+* Added support for reading from remote sources with gzip or bz2 compression 
+  (:issue:`463` - :user:`H-Max`).
+  
+* Changed keyed sorts so that comparisons are only by keys. By :user:`DiegoEPaez`, 
+  :issue:`466`.
+
+* Documentation improvements by :user:`gamesbook` (:issue:`458`).
 
 Version 1.2.0
 -------------